requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.5
django-filter==21.1
django-phonenumber-field==5.2.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
packaging==21.0
phonenumbers==8.12.33
Pillow==8.3.2
psycopg2-binary==2.8.6
PyJWT==2.3.0
pyparsing==3.0.0
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
uWSGI==2.0.20
